+++ slipp/build.gradle.kts
@@ -19,7 +19,7 @@ dependencies {
     implementation("org.springframework:spring-jdbc:$springVersion")
     implementation("org.springframework:spring-web:$springVersion")
 
-//    implementation("org.apache.commons:commons-dbcp2:2.6.0")
+    implementation("org.apache.commons:commons-dbcp2:2.6.0")
 
     runtimeOnly("com.h2database:h2:1.4.199")
     runtimeOnly("mysql:mysql-connector-java:8.0.17")
