+++ slipp/src/main/java/slipp/controller/ProfileController.java
@@ -2,6 +2,7 @@ package slipp.controller;
 
 import nextstep.jdbc.JdbcTemplate;
 import nextstep.mvc.asis.Controller;
+import slipp.dao.ConnectionManager;
 import slipp.dao.UserDao;
 import slipp.domain.User;
 
@@ -9,7 +10,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 public class ProfileController implements Controller {
-    private final UserDao userDao = new UserDao(new JdbcTemplate());
+    private final UserDao userDao = new UserDao(new JdbcTemplate(ConnectionManager.getDataSource()));
 
     @Override
     public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
