+++ slipp/src/main/java/slipp/controller/UpdateUserController.java
@@ -4,6 +4,7 @@ import nextstep.jdbc.JdbcTemplate;
 import nextstep.mvc.asis.Controller;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import slipp.dao.ConnectionManager;
 import slipp.dao.UserDao;
 import slipp.domain.User;
 import slipp.dto.UserUpdatedDto;
@@ -13,7 +14,7 @@ import javax.servlet.http.HttpServletResponse;
 
 public class UpdateUserController implements Controller {
     private static final Logger log = LoggerFactory.getLogger(UpdateUserController.class);
-    private final UserDao userDao = new UserDao(new JdbcTemplate());
+    private final UserDao userDao = new UserDao(new JdbcTemplate(ConnectionManager.getDataSource()));
 
     @Override
     public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
